feat(lua): add ISLSkillModule, ISLSkillModuleBinding types
These allow for logical grouping of Skills into "modules" which can be
translated/rotated/scaled as a unit.

This makes maintaining the position of skills much easier, as each module can
have its layout defined in a local coordinate space instead of needing to keep
in mind the global position of all other skills.