feat(skillGraph): add stats tracking to ISLSkillGraph
Optimistically I'm going to assume that I can instantiate a single graph and
that all lua scripts can subsequently pull data from it.

This may not be true, in which case we'll need to save these stats off to player
status and only recalculate from the graph when it updates.

That's not the end of the world if it comes to that.