feat(stats): stats are applied when the player loads
Previously, stats were applied every time the graph was initialized.

This is problematic because we don't want to apply the stats every time you open
the UI, instead we want to apply them when you spawn your character into a game
world and when you make changes to the graph.