doc: update README to remove FU compatability note
I'm starting to think that FU doesn't play well with others (gasp), and looking
through some of their stuff I'm not 100% that compatability is guaranteed here.

In particular, we're both modifying the player and monster event scripts to
handle damage events, and those may conflict because of how FU has decided to
implement their changes.